Modify renderSetupBoard, add selection clear/reset
Refactor the renderSetupBoard function. Now, when the player places all 10 ships on the board, the ship selection UI is cleared, and any selected ship is deselected.
Add a resetShipSelectionEl function. The function is responsible for resetting the ship selection UI by overwriting its inner HTML with the default ship options.